+++ backend/Controllers/OrderController.cs
@@ -1,7 +1,11 @@
-﻿using backend.Interfaces;
+﻿using System.Security.Claims;
+using backend.Auth.Model;
+using backend.Interfaces;
 using backend.RequestDtos.Order;
 using backend.RequestDtos.Price;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.JsonWebTokens;
 
 namespace backend.Controllers;
 
@@ -9,18 +13,53 @@ namespace backend.Controllers;
 public class OrderController : ControllerBase
 {
     private readonly IOrderService _orderService;
+    private readonly IHotelService _hotelService;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public OrderController(IOrderService orderService)
+
+    public OrderController(IOrderService orderService, IHotelService hotelService, IHttpContextAccessor httpContextAccessor)
     {
         _orderService = orderService;
+        _hotelService = hotelService;
+        _httpContextAccessor = httpContextAccessor;
     }
 
     [HttpGet]
+    [Authorize]
     public async Task<IActionResult> GetAllOrders([FromRoute] Guid hotelId)
     {
         try
         {
+            if (!_httpContextAccessor.HttpContext.User.IsInRole(Roles.Admin) && 
+                _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != await _hotelService.GetUserIdByHotel(hotelId))
+            {
+                return Forbid();
+            }
+            
             var orders = await _orderService.GetAllOrders(hotelId);
+            return Ok(orders);
+        }
+        catch (Exception)
+        {
+            return NotFound("Hotel not found");
+        }
+    }
+
+    [HttpGet]
+    [Authorize]
+    [Route("myOrders")]
+    public async Task<IActionResult> GetAllMyOrders([FromRoute] Guid hotelId)
+    {
+        try
+        {
+            if (!_httpContextAccessor.HttpContext.User.IsInRole(Roles.Admin) && 
+                _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != await _hotelService.GetUserIdByHotel(hotelId))
+            {
+                return Forbid();
+            }
+            
+            var userId = _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
+            var orders = await _orderService.UserOrders(hotelId, userId);
             return Ok(orders);
         }
         catch (Exception)
@@ -30,11 +69,19 @@ public class OrderController : ControllerBase
     }
     
     [HttpGet]
+    [Authorize]
     [Route("{id:guid}")]
     public async Task<IActionResult> GetOrderById([FromRoute] Guid hotelId, [FromRoute] Guid id)
     {
         try
         {
+            if (!_httpContextAccessor.HttpContext.User.IsInRole(Roles.Admin) && 
+                _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != await _hotelService.GetUserIdByHotel(hotelId) &&
+                _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != await _orderService.GetUserIdByOrder(id))
+            {
+                return Forbid();
+            }
+            
             var order = await _orderService.GetOrderById(hotelId, id);
             return Ok(order);
         }
@@ -49,6 +96,7 @@ public class OrderController : ControllerBase
     }
     
     [HttpPost]
+    [Authorize]
     public async Task<IActionResult> AddOrder([FromRoute] Guid hotelId, [FromBody] AddOrderRequest request)
     {
         if (!ModelState.IsValid)
@@ -63,7 +111,13 @@ public class OrderController : ControllerBase
         
         try
         {
-            var order = await _orderService.AddOrder(hotelId, request);
+            if (!_httpContextAccessor.HttpContext.User.IsInRole(Roles.Admin) && !_httpContextAccessor.HttpContext.User.IsInRole(Roles.Client))
+            {
+                return Forbid();
+            }
+            
+            var userId = _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
+            var order = await _orderService.AddOrder(hotelId, request, userId);
             return CreatedAtAction(nameof(AddOrder), new { id = order.Id }, order);
         }
         catch (Exception e)
@@ -77,6 +131,7 @@ public class OrderController : ControllerBase
     }
     
     [HttpPut]
+    [Authorize]
     [Route("{id:guid}")]
     public async Task<IActionResult> UpdateOrder([FromRoute] Guid hotelId, [FromRoute] Guid id, [FromBody] EditOrderRequest request)
     {
@@ -92,6 +147,12 @@ public class OrderController : ControllerBase
         
         try
         {
+            if (!_httpContextAccessor.HttpContext.User.IsInRole(Roles.Admin) && 
+                _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != await _orderService.GetUserIdByOrder(id))
+            {
+                return Forbid();
+            }
+            
             var order = await _orderService.UpdateOrder(hotelId, id, request);
             return Ok(order);
         }
@@ -107,11 +168,18 @@ public class OrderController : ControllerBase
     }
     
     [HttpDelete]
+    [Authorize]
     [Route("{id:guid}")]
     public async Task<IActionResult> DeleteOrder([FromRoute] Guid hotelId, [FromRoute] Guid id)
     {
         try
         {
+            if (!_httpContextAccessor.HttpContext.User.IsInRole(Roles.Admin) && 
+                _httpContextAccessor.HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != await _orderService.GetUserIdByOrder(id))
+            {
+                return Forbid();
+            }
+            
             await _orderService.DeleteOrder(hotelId, id);
             return NoContent();
         }
